before backup snapshot, checkout if backup exists make sync snapshot step stand alone

+++ scripts/backup_snapshot.py
@@ -10,6 +10,17 @@ def check_or_do(network):
     # check the snapshot is ok or not with the manifest file
     # if not, recover the backdir, then exit.
     # if ok, rm backup dir, and do the next step
+
+    # check if backup exists, delete it
+    if_backup_exists_status = os.system(
+        "kubectl exec -it -n starcoin-%s starcoin-1 -- bash -c 'if [ -d /sc-data/snapshotbak ]; then rm -rf /sc-data/snapshotbak;fi'" % network
+    )
+    if if_backup_exists_status != 0:
+        print("check if backup exists then delete failed, if_backup_exists_status:%s" %
+              if_backup_exists_status)
+        sys.exit(1)
+    print("check if backup exists then delete succeeded")
+
     backup_snapshot_status = os.system(
         "kubectl exec -it -n starcoin-%s starcoin-1 -- bash -c 'cp -r /sc-data/snapshot /sc-data/snapshotbak'" % network)
     if backup_snapshot_status != 0:
